* Fix IsSingleServer to return true when it's actually single. * Move ReplicateWrite to live only on Coordinator. * Change SequenceNumber on Point to be a uint64 to support higher numbers and host ids being the first part of it. * Move time and sequence_number assigning to the coordinator instead of in the datastore. * Roll RequestLogging and PersistentAtomicIncrement into the Datastore interface. * Add PersistentAtomicIncrement to LevelDbDatastore to support non-colliding sequence numbers.

+++ src/engine/engine.go
@@ -298,7 +298,7 @@ func (self *QueryEngine) executeCountQueryWithGroupBy(user common.User, database
 		return err
 	}
 
-	var sequenceNumber uint32 = 1
+	var sequenceNumber uint64 = 1
 	fields := []string{}
 
 	for _, aggregator := range aggregators {
